EmployeeController: make EmployeeModelAssembler implicit/internal dependency
you can say something like, but sidd, if there are multiple EmployeeControllers we will be creating multiple EmployeeModelAssemblers. is it not better to share same EmployeeModelAssembler as it's just a utility class...
and i would say definitely we should do this but when we are certain that we're going to create/need multiple EmployeeControllers. right now doing it is more of a premature optimization.
at least imo, rn...